refactor: extract honest compute core microcrate
